refactor: controller 대체
